Test to get the report names and paths into the template

+++ main.py
@@ -13,9 +13,18 @@ render_file('config.json', [data_process], storages=[])
 output_store_path = Path('reports') / f'run_{data_parameters.TODAY_STR}'
 precipy_output_path = Path('output_files')
 
-n = 1
-while output_store_path.exists():
-    output_store_path = Path('reports') / f'run_{data_parameters.TODAY_STR}_{str(n)}'
-    n = n + 1
+with open(precipy_output_path / 'data_parameters.json') as f:
+    params = json.load(f)
 
-shutil.copytree(precipy_output_path, output_store_path, shutil.copy2)
+output_store_path = params.get('ARCHIVE_REPORT_DIR')
+if output_store_path.exists():
+    print(
+f"""
+The output directory {output_store_path} already exists. 
+
+For safety reasons an archive copy will not be made. Your most recent run is available in 'output_files'.
+"""
+    )
+
+else:
+    shutil.copytree(precipy_output_path, output_store_path, shutil.copy2)
